+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Change Log
 
-## [4.51.0](https://github.com/plivo/plivo-php/tree/v4.51.1) (2023-06-28)
+## [4.52.0](https://github.com/plivo/plivo-php/tree/v4.51.1) (2023-07-18)
 -  Removed object_id and object_type in parameter as well as resonse in [list all numbers API]
+
+## [v4.51.0](https://github.com/plivo/plivo-php/tree/v4.51.0) (2023-06-28)
+**Audio Streaming**
+- API support for starting, deleting, getting streams on a live call
+- XML creation support for stream element
 
 ## [4.50.1](https://github.com/plivo/plivo-php/tree/v4.50.1) (2023-05-14)
 -  Fix warnings on [list all numbers API]